CPUFREQ: INTERACTIVE: minor code fixes

+++ drivers/cpufreq/cpufreq_interactive.c
@@ -64,7 +64,7 @@ static struct mutex gov_lock;
 
 /* Hi speed to bump to from lo speed when load burst (default max) */
 #define DEFAULT_HISPEED_FREQ 1024000
-static u64 hispeed_freq;
+static int hispeed_freq;
 
 /* Go to hi speed when CPU load at or above this value. */
 #define DEFAULT_GO_HISPEED_LOAD 99
